server_mvc/requirements_docker.txt
Flask_Cors==4.0.0
annotated-types==0.6.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
cryptography==42.0.7
dataclasses==0.6
deep-translator==1.11.4
dnspython==2.6.1
email_validator==2.1.1
filelock==3.13.1
Flask==3.0.3
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.1
safetensors==0.4.3
soupsieve==2.5
SQLAlchemy==2.0.30
sympy==1.12
tbb==2021.11.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
urllib3==2.2.1
uuid==1.30
Werkzeug==3.0.3
